--- src/main/java/com/jobsity/tenpingbowling/interfaces/BowlingBoardService.java
@@ -1,6 +1,0 @@
-package com.jobsity.tenpingbowling.interfaces;
-
-public interface BowlingBoardService {
-
-    void viewBowlingBoardResult();
-}
