--- src/main/java/com/jobsity/tenpingbowling/models/Player.java
@@ -1,14 +1,0 @@
-package com.jobsity.tenpingbowling.models;
-
-public class Player {
-
-    private String name;
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-}
