+++ src/main/java/com/tenpingbowling/Component/BowlingScoreFile.java
@@ -1,0 +1,14 @@
+package com.tenpingbowling.Component;
+
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Configuration;
+
+import java.io.File;
+
+@Configuration
+public class BowlingScoreFile extends File {
+
+    public BowlingScoreFile(@Value("${file.path}")String pathname) {
+        super(pathname);
+    }
+}
