+++ src/main/java/com/tenpingbowling/interfaces/BowlingBoardService.java
@@ -1,0 +1,6 @@
+package com.tenpingbowling.interfaces;
+
+public interface BowlingBoardService {
+
+    void viewBowlingBoardResult();
+}
