+++ src/main/java/com/tenpingbowling/interfaces/ScoreFrameCalculatorService.java
@@ -1,0 +1,12 @@
+package com.tenpingbowling.interfaces;
+
+import com.tenpingbowling.models.Frame;
+import com.tenpingbowling.models.Player;
+
+import java.util.ArrayList;
+import java.util.Map;
+
+public interface ScoreFrameCalculatorService {
+
+    Map<Player, ArrayList<Frame>> calculateFrameScore();
+}
