+++ src/main/java/com/tenpingbowling/interfaces/ScoreFrameService.java
@@ -1,0 +1,13 @@
+package com.tenpingbowling.interfaces;
+
+import com.tenpingbowling.models.Frame;
+
+import java.util.ArrayList;
+import java.util.Iterator;
+
+public interface ScoreFrameService {
+
+    ArrayList<Frame> getFrames(ArrayList<String> bowlingGameInformationByPlayerName);
+    Frame.Builder buildFrameBuilder(Iterator<String> scoreIterator, Frame.Builder frameBuilder);
+
+}
