+++ src/main/java/com/tenpingbowling/interfaces/ScoreMapService.java
@@ -1,0 +1,18 @@
+package com.tenpingbowling.interfaces;
+
+import com.tenpingbowling.models.Frame;
+import com.tenpingbowling.models.Player;
+
+import java.util.ArrayList;
+import java.util.Map;
+import java.util.Set;
+
+public interface ScoreMapService
+{
+
+    Map<Player, ArrayList<Frame>> getScoreMap();
+    Map<Player, ArrayList<Frame>> buildScoreMap();
+    void addPlayers(ArrayList<String> scoreLine);
+    void addPlayerScores(ArrayList<String> bowlingGameInformation, Set<Player> players);
+
+}
