+++ src/main/java/com/tenpingbowling/models/Player.java
@@ -1,0 +1,14 @@
+package com.tenpingbowling.models;
+
+public class Player {
+
+    private String name;
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+}
