+++ src/main/java/com/tenpingbowling/models/Roll.java
@@ -1,0 +1,30 @@
+package com.tenpingbowling.models;
+
+import com.tenpingbowling.Enums.ScoreType;
+
+public class Roll {
+
+    private String score;
+    private ScoreType rollType;
+
+    public Roll(String score, ScoreType rollType) {
+        this.score = score;
+        this.rollType = rollType;
+    }
+
+    public String getScore() {
+        return score;
+    }
+
+    public void setScore(String score) {
+        this.score = score;
+    }
+
+    public ScoreType getRollType() {
+        return rollType;
+    }
+
+    public void setRollType(ScoreType rollType) {
+        this.rollType = rollType;
+    }
+}
