+++ src/main/java/com/tenpingbowling/resources/application-prod.properties
@@ -1,0 +1,1 @@
+file.path=${file}
