+++ src/main/java/com/tenpingbowling/resources/application-test.properties
@@ -1,0 +1,1 @@
+file.path=C:\\Users\\Osvaldo\\Desktop\\score2.txt456456
